refactor: UserCoupon, UserGifticon 엔티티 패키지 이동
